Extract `classifieds::rss` module
